cosmetic changes

+++ openm/libopenm/db/omc/modelExpressionSql.cpp
@@ -451,9 +451,9 @@ const string ModelBaseExpressionSql::translateSimpleFnc(const string & i_srcMsg,
 // or:
 //  OM_VAR(acc0)
 //  =>
-//  OM_SUM((acc0 - OM_AVG(acc0)) * (acc0 - OM_AVG(acc0))) / (OM_COUNT(acc0) – 1)
+//  OM_SUM((acc0 - OM_AVG(acc0)) * (acc0 - OM_AVG(acc0))) / (OM_COUNT(acc0) â€“ 1)
 //  =>
-//  SUM((M1.acc0 - T2.ex2) * (M1.acc0 - T2.ex2)) / (COUNT(M1.acc0) – 1)
+//  SUM((M1.acc0 - T2.ex2) * (M1.acc0 - T2.ex2)) / (COUNT(M1.acc0) â€“ 1)
 //
 /** translate aggregation or non - aggregation function into sql */
 const string ModelAggregationSql::translateAggregationFnc(
@@ -975,7 +975,7 @@ bool isAccInExpr(
             bool isAcc = false;
             size_t nLen = 0;
             int accPos;
-            for (accPos = 0; accPos < nAcc; accPos++) {
+            for (accPos = 0; accPos < (int)nAcc; accPos++) {
 
                 nLen = i_accNames[accPos].length();
                 isAcc = equalNoCase(i_accNames[accPos].c_str(), i_expr.c_str() + nPos, nLen);
